[DAD-1061] refactor: count api 코드 리팩토링
